Switch cursor commands over to be ContextualCommands
It is more consistent with the rest of the commands and allows us to remove any remaining remnants of the old pipeline commands.
The "pipeline" commands were based on the idea of WPF routed commands, but it seemed as if that introduced more complexity without much benefit.  In the future, we may wish to revisit something like RoutedCommands, but for now, this is better.