Remove TextRight implementation of IEnumerable
It's not really a collection in the "classical" sense, and it made testing equivalence in a pain in the ass.  The latter is not the best reason to reason to make the change, but it indicates that it didn't really "look" like a collection.
FYI, it was a pain because it made FluentAssertions treat the object as a collection, not as an object, so other properties (like heading-levels, for instance) would not be compared properly.